backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2

# PDF processing
PyMuPDF==1.23.18

# AI/LLM
openai==1.12.0

# Security
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# PDF report generation
fpdf2==2.7.8

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
pytest-cov==4.1.0
